refactor: position components

+++ Rogueskiv.Core/Components/Walls/HorizontalWallComp.cs
@@ -5,8 +5,8 @@
         public HorizontalWallComp(int x, int y, int width)
             : base(x, y, width) { }
 
-        protected override int FixedPosition => Y;
-        protected override int VariablePosition => X;
+        protected override float FixedPosition => Y;
+        protected override float VariablePosition => X;
 
         protected override float GetFixedPosition(PositionComp position) => position.Y;
         protected override float GetVariablePosition(PositionComp position) => position.X;
