+++ Rogueskiv.Core/Components/Walls/VerticalWallComp.cs
@@ -5,8 +5,8 @@
         public VerticalWallComp(int x, int y, int height)
             : base(x, y, height) { }
 
-        protected override int FixedPosition => X;
-        protected override int VariablePosition => Y;
+        protected override float FixedPosition => X;
+        protected override float VariablePosition => Y;
 
         protected override float GetFixedPosition(PositionComp position) => position.X;
         protected override float GetVariablePosition(PositionComp position) => position.Y;
